BenchmarkToken partially works
