+++ contracts/benchmark/token-in-progress.js
@@ -1,18 +1,18 @@
-import { Address, State, Verify, Uint64 } from "orbs-contract-sdk/v1";
+import { Address, State, Verify, Uint32 } from "orbs-contract-sdk/v1";
 
-// Define uint64 later
-const TOTAL_SUPPLY = Uint64("10000000000");
+// Define uint32 later
+const TOTAL_SUPPLY = Uint32("10000000");
 
 // doesn't do anything except validation
 export function TransferEvent(from, to, amount) {
     Verify.bytes(from);
     Verify.bytes(to);
-    Verify.uint64(amount);
+    Verify.uint32(amount);
 }
 
 export function _init() {
     const ownerAddress = Address.getSignerAddress();
-    State.writeUint64(ownerAddress, TOTAL_SUPPLY);
+    State.writeUint32(ownerAddress, TOTAL_SUPPLY);
 }
 
 export function totalSupply() {
@@ -20,21 +20,21 @@ export function totalSupply() {
 }
 
 export function transfer(amount, targetAddress) {
-    Verify.uint64(amount);
+    Verify.uint32(amount);
     Verify.bytes(targetAddress);
 
     // sender
     const callerAddress = Address.getCallerAddress();
-    const callerBalance = State.readUint64(callerAddress);
+    const callerBalance = State.readUint32(callerAddress);
     if (callerBalance < amount) {
         throw new Error(`transfer of ${amount} failed since balance is only ${callerBalance}`);
     }
-    State.writeUint64(callerAddress, callerBalance-amount);
+    State.writeUint32(callerAddress, callerBalance-amount);
 
     // recipient
     Address.validateAddress(targetAddress);
-    const targetBalance = State.readUint64(targetAddress);
-    State.writeUint64(targetAddress, targetBalance+amount);
+    const targetBalance = State.readUint32(targetAddress);
+    State.writeUint32(targetAddress, targetBalance+amount);
 
     Events.emitEvent(TransferEvent, from, to, amount);
 }
@@ -42,5 +42,5 @@ export function transfer(amount, targetAddress) {
 export function getBalance() {
     Verify.Bytes(targetAddress);
     Address.validateAddress(targetAddress);
-    return State.readUint64(targetAddress);
+    return State.readUint32(targetAddress);
 }
